Refactor: swagger annotation 분리
